SFB-65: move the simple linter extension also to a separate file
